So, basically, the reason I didn't commit
at all during this assignment, was that I couldn't get it to work with Visual Studio.  I reinstalled 3 times, and tried to revert the extension and all the jazz, and it was overall just wasting time.  So I coded the Assignment instead.  I didn't want to lose marks on not being able to hand in anything if I was trying to get it to work. 

Also, for not handing it in on time, you told me in one of your c++ labs that I could hand it in the next morning.

Oh, also, that one room in NE1, the corner chair at the back is broken :p haha, goodnight Chris.